Surfacic lights
Add surfacic lights

+++ Synthese2/Color.cpp
@@ -43,7 +43,7 @@ Color Color::operator*(const Color& other) const
 }
 Color Color::operator*(const double other) const
 {
-    return Color{static_cast<int>(m_r * other), static_cast<int>(m_g * other), static_cast<int>(m_b * other)};
+    return Color{static_cast<double>(m_r * other), static_cast<double>(m_g * other), static_cast<double>(m_b * other)};
 }
 bool Color::operator==(const Color &other) const {
     return m_r == other.m_r && m_g == other.m_g && m_b == other.m_b;
